# public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Hotel Management API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/post/login": {
            "post": {
                "summary": "User login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "rememberMe": {
                                        "type": "boolean"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Password is wrong"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/api/feedback/softdelete/{id}": {
            "delete": {
                "summary": "Soft delete a feedback (admin only)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The ID of the feedback to be soft deleted"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feedback soft deleted successfully"
                    },
                    "400": {
                        "description": "Feedback already soft deleted or validation error"
                    },
                    "401": {
                        "description": "Invalid token"
                    },
                    "403": {
                        "description": "User is not admin"
                    },
                    "404": {
                        "description": "Feedback or token not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/feedback/post": {
            "post": {
                "summary": "Feedback post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "fromMail": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "fromMail",
                                    "description"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Feedback create successfully"
                    },
                    "400": {
                        "description": "Email is required or Please enter a valid email or Feedback description is required or Feedback description maximum 100 characters."
                    }
                }
            }
        },
...
    }
}